Updated code as per Github comments
