+++ src/types/wallet.ts
@@ -1,0 +1,6 @@
+// Supported wallet types
+export enum WalletType {
+    EVM = "evm",
+    SOLANA = "solana",
+    BITCOIN = "bitcoin",
+}
