+++ src/utils/checkEmpty.ts
@@ -1,0 +1,2 @@
+// Function to check if the given value is empty or not
+export const checkEmpty = (value: string): boolean => !value || value.trim().length === 0;
