+++ src/validation/address.ts
@@ -1,0 +1,19 @@
+// Import necessary modules and types
+import { WalletType } from "../types/wallet";
+
+// Define regexes Bitcoin addresses
+const p2pkh = /^[13][a-km-zA-HJ-NP-Z0-9]{25,34}$/;
+const p2sh = /^3[a-km-zA-HJ-NP-Z0-9]{25,34}$/;
+const bech32 = /^(bc1)[0-9a-z]{39,59}$/;
+
+// Combine the above Bitcoin regexes into a single regex
+export const regexBitcoin = new RegExp(`${p2pkh.source}|${p2sh.source}|${bech32.source}`);
+
+// Define regex for Solana addresses
+export const regexSolana = new RegExp(/^[1-9A-HJ-NP-Za-km-z]{32,44}$/);
+
+// Create a partial record that maps wallet types to their corresponding regex patterns
+export const allWalletRegexes: Partial<Record<WalletType, RegExp>> = {
+    [WalletType.SOLANA]: regexSolana,
+    [WalletType.BITCOIN]: regexBitcoin,
+};
